Adds and uses modelNameCache
